[#2332] Move IVC folding expresion evaluation into separate module
